pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "comm_centralizer"
version = "0.1.0"
description = "Communication centralizer for multiple platforms"
requires-python = ">=3.8"
dependencies = [
    # Core requirements
    "python-dotenv>=0.19.0",
    "openai>=1.0.0",
    "requests>=2.25.1",

    # API integrations
    "google-api-python-client>=2.33.0",
    "google-auth-httplib2>=0.1.0",
    "google-auth-oauthlib>=0.5.0",
    "slack-sdk>=3.15.2",
    "discord.py>=2.0.0",
    "linkedin-api>=2.0.0",

    # Data processing
    "pandas>=1.3.5",
    "numpy>=1.21.5",

    # Web automation
    "selenium>=4.1.0",
    "webdriver-manager>=3.5.2",

    # API clients
    "gspread>=5.0.0",
    "Calendly-Python>=0.1.0",

    # NLP processing
    "spacy>=3.4.0",
    "spacy-llm>=0.6.0",
    "spacy-transformers>=1.1.0",

    # Job scheduling
    "python-crontab>=3.2.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=7.0.0",
    "pytest-cov>=2.12.1",
    "pytest-asyncio>=0.21.0",
    "colorama>=0.4.4",

    # Development tools
    "pre-commit>=2.17.0",
    "black>=22.1.0",
    "flake8>=4.0.1",
    "mypy>=0.931",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]
